pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skill-auditor"
version = "0.1.0"
description = "Static analysis and safety-scoring tool for LLM agent skills (like npm audit for AI tools)"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
    "PyYAML>=6.0",
]

[tool.setuptools.packages.find]
include = ["skill_auditor*", "ingestion*", "ranking*"]

[project.scripts]
skill-audit = "skill_auditor.cli:main"
skill-sync = "ingestion.sync:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]

[tool.pytest.ini_options]
testpaths = ["tests", "evaluation"]
addopts = "-v"
